feat: extend color variants to include text color
